requirements.txt
# 核心库
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0

# 模型库
transformers>=4.30.0
scikit-learn>=1.3.0

# 数据处理
pandas>=2.0.0
pydub>=0.25.0
pyyaml>=6.0

# 可视化与分析
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.0.0

# LLM API
openai>=1.0.0
requests>=2.28.0

# 工具
tqdm>=4.65.0
pydantic>=2.0.0

# 开发工具
ipython>=8.0.0
jupyter>=1.0.0
black>=23.0.0